requirements/base.txt
Django==1.9.6
bokeh==0.11.1
fuzzywuzzy==0.10.0
pandas==0.18.1
psycopg2==2.6.1
python-dateutil==2.5.3
python-Levenshtein==0.12.0
requests==2.10.0
raven==5.17.0
django-extended-choices==1.0.7
django-moj-template==0.23.1

#Irat healthcheck and ping package
git+https://github.com/ministryofjustice/django-moj-irat.git@4d54b86b1cb574fe787ba0fb8a992cf352f8eba6
